[linux] Add uaccess
Add user access API for linux.

On linux userspace virtual == user == phys addresses.
Physical addresses also being the same is wrong, but there is no general way
of converting userspace addresses to physical as what appears to be
contiguous in userspace is physically fragmented.
Currently only the DMA memory is special-cased, but it's conversion to bus
addresses is done in phys_to_bus.
This is known to break virtio as it is passing phys addresses to the virtual
device.

Signed-off-by: Piotr Jaroszyński <p.jaroszynski@gmail.com>